fix: cast AlphaEdit solve operands to double before matmul, not after
P matrix is float32, layer_ks may be float16 (model loaded in half precision).
(P_i @ layer_ks).double() crashes because the matmul itself fails on mixed
dtypes. Now all operands are cast to double upfront.

+++ src/algorithms/alphaedit_with_hooks.py
@@ -117,21 +117,23 @@ def apply_alphaedit_with_hooks(
         # Hooks provide the INNER term only. AlphaEdit always wraps with:
         #   lhs = P @ (inner + cache_c) + L2*I
         # This ensures the LHS is full-rank even when the inner term is low-rank.
-        P_i = P[i, :, :].cuda()
-        cache_c_i = cache_c[i, :, :].cuda() if cache_c is not None else torch.zeros(
-            layer_ks.shape[0], layer_ks.shape[0], device="cuda")
-        L2_eye = hparams.L2 * torch.eye(layer_ks.shape[0], device="cuda")
+        # Cast everything to double upfront — P is float32, layer_ks may be float16
+        P_i = P[i, :, :].cuda().double()
+        cache_c_i = cache_c[i, :, :].cuda().double() if cache_c is not None else torch.zeros(
+            layer_ks.shape[0], layer_ks.shape[0], device="cuda", dtype=torch.float64)
+        L2_eye = hparams.L2 * torch.eye(layer_ks.shape[0], device="cuda", dtype=torch.float64)
 
         if hooks.build_lhs is not None:
             inner = hooks.build_lhs(layer, layer_ks, None, hparams, state)
         else:
-            inner = layer_ks @ layer_ks.T
+            inner = layer_ks.double() @ layer_ks.double().T
 
-        lhs = P_i @ (inner + cache_c_i) + L2_eye
+        lhs = P_i @ (inner.double() + cache_c_i) + L2_eye
 
         # AlphaEdit solve: lhs @ upd = P @ K @ resid^T
+        # Cast to double BEFORE matmul — P is float32, layer_ks may be float16
         upd_matrix = torch.linalg.solve(
-            lhs.double(), (P_i @ layer_ks @ resid.T).double()
+            lhs.double(), P_i.double() @ layer_ks.double() @ resid.double().T
         )
 
         weight_name = f"{hparams.rewrite_module_tmp.format(layer)}.weight"
